:zap: Add condition to not re-add ssh key

+++ scripts/generate-ssh.key.sh
@@ -3,13 +3,15 @@
 GREEN='\e[32m'
 NC='\e[0m'
 
-echo -e "\n${GREEN}Setting Up SSH config...${NC}" &&
-	ssh-keygen -t ed25519 -C "muhammadkhan1607@gmail.com" -f ~/.ssh/id-github -N "" &&
-	eval "$(ssh-agent -s)" &&
-	ssh-add ~/.ssh/id-github &&
-	echo -e -n "${GREY}Enter a title for your new github ssh key:${NC} " &&
-	read -r ssh_key_title &&
-	gh ssh-key add ~/.ssh/id-github.pub -t "${ssh_key_title}"
-  gh ssh-key add ~/.ssh/id-github.pub --type signing
+if [ ! -f "$HOME/.ssh/id-github" ]; then
+  echo -e "\n${GREEN}Setting Up SSH config...${NC}" &&
+    ssh-keygen -t ed25519 -C "muhammadkhan1607@gmail.com" -f ~/.ssh/id-github -N "" &&
+    eval "$(ssh-agent -s)" &&
+    ssh-add ~/.ssh/id-github &&
+    echo -e -n "${GREY}Enter a title for your new github ssh key:${NC} " &&
+    read -r ssh_key_title &&
+    gh ssh-key add ~/.ssh/id-github.pub -t "${ssh_key_title}"
+    gh ssh-key add ~/.ssh/id-github.pub --type signing
 
-[ ! -f "$HOME/.ssh/config" ] && cp ~/dotfiles/config/.ssh/* ~/.ssh/
+  [ ! -f "$HOME/.ssh/config" ] && cp ~/dotfiles/config/.ssh/* ~/.ssh/
+fi
